requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.15
urllib3==1.26.4
Werkzeug==2.0.1
pandas==1.2.4
wheel==0.36.2